UtilityMethodTestCaseTest: split off the test for the `expectPhpcsException()` method
The `UtilityMethodTestCase::expectPhpcsException()` method provides a polyfill for changed PHPUnit functionality. To allow for testing this polyfill properly, it should be tested without the PHPUnit Polyfills active as external standards using the `UtilityMethodTestCase` may not implement the PHPUnit Polyfills.

To do so, the tests specific to the `UtilityMethodTestCase::expectPhpcsException()` method have been split off into their own test class which extends the `UtilityMethodTestCase` directly, instead of extending the `PolyfilledTestCase`.
